% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAccuracy.R
\name{getAccuracy}
\alias{getAccuracy}
\title{Calculate the Extended Accuracy from a Confusion Matrix}
\usage{
getAccuracy(table, n_of_parallel_diagonals_to_consider = 0)
}
\arguments{
\item{table}{A square matrix representing the confusion matrix data.
Each element should be an integer count of predictions versus
actual class labels. The matrix must be square.}

\item{n_of_parallel_diagonals_to_consider}{An integer indicating how many
diagonals parallel to the main diagonal should be considered in the
accuracy calculation. This allows for a tolerance in classification
accuracy where predictions close to the correct class are counted as
correct. This value must be non-negative and less than the dimension of
the matrix.}
}
\value{
A list containing two elements:
        - `accuracy`: The extended accuracy as a decimal, representing the
          probability that the classification is correct or nearly correct
          within the specified tolerance.
        - `expected_accuracy`: The expected accuracy assuming random classification,
          adjusted for the number of diagonals considered.
}
\description{
This function computes the extended accuracy of a classification model by considering
not only the main diagonal (perfect classification matches) but also adjacent
diagonals if specified. This allows for a measure of accuracy that can accept
near-miss classifications as correct within a certain tolerance.
}
\examples{
table <- matrix(c(10, 2, 1, 2, 25, 3, 4, 0, 18), nrow = 3, byrow = TRUE)
getAccuracy(table, 1)  # Includes main diagonal and one parallel on each side
}
