% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConfusionMatrix.R
\name{plotConfusionMatrix}
\alias{plotConfusionMatrix}
\title{Plot a Confusion Matrix with Customizable Options}
\usage{
plotConfusionMatrix(
  table,
  color = c("grey_scaled", "rgb_scaled"),
  rgb_code = c(0, 0, 1),
  horizontal_label = "horizontal_label",
  vertical_label = "vertical_label",
  class_labels,
  label_half_width = 0.5,
  label_offset_factor = 2,
  horizontal_label_srt = 0,
  horizontal_label_cex = 1,
  vertical_label_srt = 90,
  vertical_label_cex = 1,
  left_margin = 5,
  top_margin = 5,
  main_diagonal_highlighting = c("framebox", "color", "both", "none"),
  rgb_code_highlighting = c(1, 0, 0),
  rgb_code_highlighting_framebox = c(0, 0, 0),
  highlighting_framebox_lwd = 1,
  n_of_parallel_diagonals_to_highlight = 0,
  percentage_display = FALSE,
  percentage_round_digits = 1
)
}
\arguments{
\item{table}{A square matrix containing confusion matrix data where each
element represents the count of predictions for each actual class.
This matrix must be square and contain integer values only.}

\item{color}{A character string specifying the color scheme for the matrix.
Options are "grey_scaled" for grayscale and "rgb_scaled" for RGB
color scaling. Default is c("grey_scaled", "rgb_scaled").}

\item{rgb_code}{A numeric vector of three elements (between 0 and 1) indicating
the RGB values for color scaling when 'color' is set to "rgb_scaled".
Default is c(0, 0, 1).}

\item{horizontal_label}{A string label for the horizontal axis. Default is
"horizontal_label".}

\item{vertical_label}{A string label for the vertical axis. Default is
"vertical_label".}

\item{class_labels}{A vector of character strings that labels the classes
corresponding to the rows and columns of the matrix.}

\item{label_half_width}{Numeric value specifying half the width of the labels.
Default is 0.5.}

\item{label_offset_factor}{Numeric value used as a multiplier to adjust label
offsets. Default is 2.0.}

\item{horizontal_label_srt}{The rotation angle for horizontal labels in degrees.
Default is 0.}

\item{horizontal_label_cex}{Numeric value specifying the character expansion
factor for horizontal labels. Default is 1.0.}

\item{vertical_label_srt}{The rotation angle for vertical labels in degrees.
Default is 90.}

\item{vertical_label_cex}{Numeric value specifying the character expansion
factor for vertical labels. Default is 1.0.}

\item{left_margin}{Numeric value specifying the left margin size of the plot.
Default is 5.0.}

\item{top_margin}{Numeric value specifying the top margin size of the plot.
Default is 5.0.}

\item{main_diagonal_highlighting}{Character vector specifying how to highlight
the main diagonal. Options are "framebox",
"color", "both", or "none". Default is c("framebox",
"color", "both", "none").}

\item{rgb_code_highlighting}{Numeric vector of three elements for the RGB values
used in highlighting. Default is c(1, 0, 0).}

\item{rgb_code_highlighting_framebox}{Numeric vector of three elements for the
RGB frame box color used in highlighting.
Default is c(0, 0, 0).}

\item{highlighting_framebox_lwd}{Numeric specifying the line width of the frame
box used in highlighting. Default is 1.0.}

\item{n_of_parallel_diagonals_to_highlight}{Integer indicating how many
diagonals parallel to the main diagonal should be highlighted.
Default is 0.}

\item{percentage_display}{Logical indicating whether to display percentages
instead of absolute numbers. Default is FALSE.}

\item{percentage_round_digits}{Integer specifying the number of decimal places
for rounding the displayed percentages.
Default is 1.}
}
\description{
This function visualizes a confusion matrix with various customization options
including color schemes, labeling, and highlighting features. It allows for
visual emphasis on specific aspects of the data, such as the main diagonal or
certain off-diagonal elements, and can represent data in absolute or percentage
terms.
}
\examples{
table <- matrix(
   c(10, 2, 1, 2, 15, 3, 4, 0, 18),
   nrow = 3,
   byrow = TRUE,
   dimnames = list(observed = c("class_1", "class_2", "class_3"),
                   predicted = c("class_1", "class_2", "class_3"))
)
plotConfusionMatrix(table, color = "rgb_scaled", rgb_code = c(1, 0.5, 0),
                    horizontal_label = "Predicted Class",
                    vertical_label = "Actual Class",
                    main_diagonal_highlighting = "color",
                    rgb_code_highlighting = c(0.5, 0.5, 0.5),
                    percentage_display = FALSE)
}
